requirements.txt
pandas==2.2.3
uvicorn==0.34.0
fastapi==0.115.8
pydantic==2.10.6
sqlalchemy==2.0.38
passlib[bcrypt]==1.7.4
python-jose==3.3.0
bcrypt==4.2.1
pyjwt==2.10.1
